+++ crates/mcp-server/examples/common/counter.rs
@@ -177,4 +177,3 @@ impl mcp_server::Router for CounterRouter {
         })
     }
 }
-
