+++ crates/mcp-server/examples/common/mod.rs
@@ -1,0 +1,2 @@
+pub mod counter;
+pub mod jsonrpc_frame_codec;
